Completely removed necessity to change Code in NthCompartmentModel.py (formerly known as SIR_Model_Network.py). Every parameter is now read from the csv files.
Introduced method to call functions inside the replacement of the placeholders. Currently not able to pass arguments, coming soon.

Renamed
- SIR_Model_Network.py to NthCompartmentModel.py
- siressentials.py to ModelEssentials.py
