+++ AdjacencyMuenster.csv
@@ -1,4 +1,4 @@
-from/to,Muenster,Steinfurt,Coesfeld,Warendorf,Borken,Recklinghausen,Unna,Hamm,Dortmund,Osnabrueck (LK),Osnabrueck (SK)
+from/to,Muenster,Steinfurt,Coesfeld,Warendorf,Borken,Recklinghausen,UnIna,Hamm,Dortmund,Osnabrueck (LK),Osnabrueck (SK)
 Muenster,0,5789,4032,4365,891,809,858,989,1176,0,749
 Steinfurt,22183,0,1952,1645,3891,0,0,0,0,5693,9859
 Coesfeld,18890,2119,0,1020,5067,4024,2944,663,1885,0,0
