+++ Time_Dependencies.py
@@ -1,0 +1,6 @@
+def none():
+    return None
+
+
+def infection_rate():
+    return 0.35
